fix third part of eslint errors
