+++ src/components/animations/RotationArrow.js
@@ -1,6 +1,12 @@
 import PropTypes from 'prop-types';
-import React, { Clock, Component, spring, SpringUtils, Value } from 'react';
-import Animated from 'react-native-reanimated';
+import React from 'react';
+import Animated, {
+  Clock,
+  Component,
+  spring,
+  SpringUtils,
+  Value,
+} from 'react-native-reanimated';
 import { interpolate } from './procs';
 
 const {
