+++ src/components/avatar-builder/Categories.js
@@ -1,0 +1,52 @@
+/* eslint-disable sort-keys */
+
+export const Categories = {
+  people: {
+    icon: 'emojiSmileys',
+    index: 0,
+    name: 'Smileys & People',
+    width: 138,
+  },
+  nature: {
+    icon: 'emojiAnimals',
+    index: 1,
+    name: 'Animals & Nature',
+    width: 143,
+  },
+  food: {
+    icon: 'emojiFood',
+    index: 2,
+    name: 'Food & Drink',
+    width: 109,
+  },
+  activities: {
+    icon: 'emojiActivities',
+    index: 3,
+    name: 'Activities',
+    width: 87,
+  },
+  places: {
+    icon: 'emojiTravel',
+    index: 4,
+    name: 'Travel & Places',
+    width: 132,
+  },
+  objects: {
+    icon: 'emojiObjects',
+    index: 5,
+    name: 'Objects',
+    width: 75,
+  },
+  icons: {
+    icon: 'emojiSymbols',
+    index: 6,
+    name: 'Symbols',
+    width: 79,
+  },
+  flags: {
+    icon: 'emojiFlags',
+    index: 7,
+    name: 'Flags',
+    width: 57,
+  },
+};
