+++ src/components/avatar-builder/EmojiSelector.js
@@ -31,60 +31,10 @@ import 'string.fromcodepoint';
 import EmojiTabBarShadow from '../../assets/emojiTabBarShadow.png';
 import { colors, fonts, position } from '../../styles';
 import { deviceUtils } from '../../utils';
+import { Categories } from './Categories';
 import TabBar from './TabBar';
 
 // TODO width attribute is temporary solution that will be removed as soon as I figure out why proper scaling does not work
-
-export const Categories = {
-  people: {
-    icon: 'emojiSmileys',
-    index: 0,
-    name: 'Smileys & People',
-    width: 138,
-  },
-  nature: {
-    icon: 'emojiAnimals',
-    index: 1,
-    name: 'Animals & Nature',
-    width: 143,
-  },
-  food: {
-    icon: 'emojiFood',
-    index: 2,
-    name: 'Food & Drink',
-    width: 109,
-  },
-  activities: {
-    icon: 'emojiActivities',
-    index: 3,
-    name: 'Activities',
-    width: 87,
-  },
-  places: {
-    icon: 'emojiTravel',
-    index: 4,
-    name: 'Travel & Places',
-    width: 132,
-  },
-  objects: {
-    icon: 'emojiObjects',
-    index: 5,
-    name: 'Objects',
-    width: 75,
-  },
-  icons: {
-    icon: 'emojiSymbols',
-    index: 6,
-    name: 'Symbols',
-    width: 79,
-  },
-  flags: {
-    icon: 'emojiFlags',
-    index: 7,
-    name: 'Flags',
-    width: 57,
-  },
-};
 
 const charFromUtf16 = utf16 =>
   String.fromCodePoint(...utf16.split('-').map(u => '0x' + u));
@@ -276,6 +226,7 @@ export default class EmojiSelector extends PureComponent {
         emojiCategory[1].offset = offset;
         allEmojiList = allEmojiList.concat(emojiCategory);
       }
+      return true;
     });
 
     allEmojiList.push({ overlay: true });
