+++ src/components/avatar-builder/TabBar.js
@@ -3,7 +3,7 @@ import LinearGradient from 'react-native-linear-gradient';
 import { colors, position } from '../../styles';
 import { ButtonPressAnimation } from '../animations';
 import { Icon } from '../icons';
-import { Categories } from './EmojiSelector';
+import { Categories } from './Categories';
 
 const TabBar = ({ categoryKeys, activeCategory, onPress }) => {
   return categoryKeys.map(c => {
